requirements.txt
streamlit
google-generativeai
python-dotenv
PyPDF2

# Audio recording and processing
streamlit-mic-recorder
streamlit-webrtc
av
pydub
librosa
soundfile
ffmpeg-python

# ML and NLP
torch
torchaudio
transformers
datasets
numpy
